connecting to frontend

+++ src/videos/videos.service.ts
@@ -3,6 +3,8 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Video } from '../videos/videos.entity';
+import { Lesson } from '../lessons/lesson.entity';
+
 
 @Injectable()
 export class VideosService {
@@ -22,16 +24,40 @@ export class VideosService {
 
   // Fetch all videos
   async findAll(): Promise<Video[]> {
-    return this.videosRepository.find();
+    return this.videosRepository.find({
+      select: ['id', 'title', 'url', 'lesson'], // Explicitly include lessonId
+      relations: ['lesson'],
+    });
   }
-
+  
   // Create a new video
   async create(video: Partial<Video>): Promise<Video> {
-    return this.videosRepository.save(video);
+    if (!video.lesson || !video.lesson.id) {
+      throw new Error('Lesson ID is required.');
+    }
+  
+    const lessonRepository = this.videosRepository.manager.getRepository(Lesson);
+    const lesson = await lessonRepository.findOne({ where: { id: video.lesson.id } });
+  
+    if (!lesson) {
+      throw new Error('Associated Lesson not found.');
+    }
+  
+    const newVideo = this.videosRepository.create({
+      ...video,
+      lesson: lesson,
+    });
+  
+    return await this.videosRepository.save(newVideo);
   }
 
   // Delete a video by ID
   async delete(id: number): Promise<void> {
+    const video = await this.videosRepository.findOne({ where: { id } });
+    if (!video) {
+      throw new Error('Video not found');
+    }
     await this.videosRepository.delete(id);
   }
+  
 }
